requirements.txt
aiocsv==1.2.3
aiohttp==3.8.3
influxdb-client==v1.35.0
confluent_kafka==1.9.2
requests==v2.28.1
prometheus-client==v0.9.0
protobuf==3.20.3
paho-mqtt==1.5.0
websocket-client==v1.3.2
